refactor: sections 폴더 생성
